requirements.txt
json5>=0.9.14
pandas>=1.4.3
Requests>=2.32.5
gspread>=5.7.2
oauth2client>=4.1.3
plotly-express>=0.4.1
